requirements.txt
streamlit==1.32.0
tensorflow-cpu==2.16.2; python_version < "3.13"
openai==1.12.0
httpx==0.27.2
Pillow==10.2.0
numpy==1.26.4
python-dotenv==1.0.1
